requirements.txt
ccxt
pandas
pymongo
scikit-learn
python-dotenv
python-binance

# models
google-genai
openai
fastapi
uvicorn
binance